revise slice size logic to sensible default with min value

+++ gcsfast/cli/upload.py
@@ -61,7 +61,7 @@ def upload_command(threads: int, slice_size: int, io_buffer: int,
         input_stream = open(file_path, "rb")
 
     upload_slice_size = slice_size if slice_size else max(
-        [stat(file_path).st_size / threads, 16 * 2**20])
+        [int(stat(file_path).st_size / threads), 16 * 2**20])
 
     executor = BoundedThreadPoolExecutor(max_workers=threads,
                                          queue_size=int(threads * 1.5))
